Revert "Fix Vptr.disp size to fit MemoryRef.disp"
This reverts commit 9dfbd2d8c51477499a681af4dd238f48e0b04f22.

+++ hphp/runtime/vm/jit/vasm-reg.h
@@ -265,7 +265,7 @@ struct Vptr {
   Vreg64 index;     // optional
   uint8_t scale;    // 1,2,4,8
   Segment seg{DS};  // DS, FS or GS
-  intptr_t disp;    // big enough to fit a MemoryRef.disp type
+  int32_t disp;
 };
 
 Vptr operator+(Vptr lhs, int32_t d);
